Attempt to add P010 support to Y4MEncoder in examples
The idea is that if the input image is a 10-bit 4:2:0 HEIF image and the output is .y4m,
output a P010 encoded file instead of an 8-bit YUV 4:2:0 file. This file
is then supposed to be fed into libultrahdr to create a jpg gain map
image.

Right now, it does result in a proper JPG gain-mapped image that is much
brighter when HDR is enabled and viewed in Google Chrome.

Signed-off-by: Eric Joyner <erj@erj.cc>